address review comments, add all arkworks crates as features
