+++ openzl-crypto/src/hash.rs
@@ -1,7 +1,5 @@
 //! Hash Functions
 
-use openzl_util::derivative;
-
 /// Hash Function
 pub trait HashFunction<COM = ()> {
     /// Input Type
@@ -27,21 +25,6 @@ where
     }
 }
 
-/// Binary Hash Function
-pub trait BinaryHashFunction<COM = ()> {
-    /// Left Input Type
-    type Left: ?Sized;
-
-    /// Right Input Type
-    type Right: ?Sized;
-
-    /// Output Type
-    type Output;
-
-    /// Computes the hash over `lhs` and `rhs`.
-    fn hash(&self, lhs: &Self::Left, rhs: &Self::Right, compiler: &mut COM) -> Self::Output;
-}
-
 /// Array Hash Function
 pub trait ArrayHashFunction<const ARITY: usize, COM = ()> {
     /// Input Type
@@ -52,71 +35,6 @@ pub trait ArrayHashFunction<const ARITY: usize, COM = ()> {
 
     /// Computes the hash over `input`.
     fn hash(&self, input: [&Self::Input; ARITY], compiler: &mut COM) -> Self::Output;
-}
-
-/// Array Hashing Utilities
-pub mod array {
-    use super::*;
-    use core::marker::PhantomData;
-
-    #[cfg(feature = "serde")]
-    use openzl_util::serde::{Deserialize, Serialize};
-
-    /// Converts `hasher` from an [`ArrayHashFunction`] into a [`HashFunction`].
-    #[inline]
-    pub fn as_unary<H, COM>(hasher: H) -> AsUnary<H, COM>
-    where
-        H: ArrayHashFunction<1, COM>,
-    {
-        AsUnary::new(hasher)
-    }
-
-    /// Unary Hash Function Converter
-    #[cfg_attr(
-        feature = "serde",
-        derive(Deserialize, Serialize),
-        serde(crate = "openzl_util::serde", deny_unknown_fields)
-    )]
-    #[derive(derivative::Derivative)]
-    #[derivative(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-    pub struct AsUnary<H, COM = ()>
-    where
-        H: ArrayHashFunction<1, COM>,
-    {
-        /// Array Hasher
-        hasher: H,
-
-        /// Type Parameter Marker
-        __: PhantomData<COM>,
-    }
-
-    impl<H, COM> AsUnary<H, COM>
-    where
-        H: ArrayHashFunction<1, COM>,
-    {
-        /// Builds a new [`HashFunction`] implementation out of an [`ArrayHashFunction`]
-        /// implementation `hasher`.
-        #[inline]
-        pub fn new(hasher: H) -> Self {
-            Self {
-                hasher,
-                __: PhantomData,
-            }
-        }
-    }
-
-    impl<H, COM> HashFunction<COM> for AsUnary<H, COM>
-    where
-        H: ArrayHashFunction<1, COM>,
-    {
-        type Input = H::Input;
-        type Output = H::Output;
-
-        #[inline]
-        fn hash(&self, input: &Self::Input, compiler: &mut COM) -> Self::Output {
-            self.hasher.hash([input], compiler)
-        }
-    }
 }
 
 /// Security Assumptions
