--- openzl-derive/README.md
@@ -1,3 +1,0 @@
-# `eclair`: **E**mbedded **C**ircuit **L**anguage **A**nd **I**ntermediate **R**epresentation
-
-
