+++ plugins/arkworks/src/rand.rs
@@ -1,24 +1,28 @@
 //! Arkworks Random Sampling Implementations
 
-use crate::{
-    ec::{
-        models::{
-            short_weierstrass_jacobian, twisted_edwards_extended, SWModelParameters,
-            TEModelParameters,
-        },
-        AffineCurve, ProjectiveCurve,
+#[cfg(feature = "ec")]
+use crate::ec::{
+    models::{
+        short_weierstrass_jacobian, twisted_edwards_extended, SWModelParameters, TEModelParameters,
     },
-    ff::{Fp256, Fp320, Fp384, Fp448, Fp64, Fp768, Fp832, UniformRand},
+    AffineCurve, ProjectiveCurve,
 };
-use openzl_util::rand::{RngCore, Sample};
+
+#[cfg(feature = "ff")]
+use crate::ff::{Fp256, Fp320, Fp384, Fp448, Fp64, Fp768, Fp832, UniformRand};
+
+pub use openzl_util::rand::*;
 
 /// Standard Distribution
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Standard;
 
 /// Builds a [`Sample`] implementation for `$projective` and `$affine` curves over the `$P` model.
+#[cfg(feature = "ec")]
 macro_rules! sample_curve {
     ($P:tt, $trait:tt, $projective:path, $affine:path $(,)?) => {
+        #[cfg(feature = "ec")]
+        #[cfg_attr(doc_cfg, doc(cfg(feature = "ec")))]
         impl<$P> Sample<Standard> for $projective
         where
             $P: $trait,
@@ -33,6 +37,8 @@ macro_rules! sample_curve {
             }
         }
 
+        #[cfg(feature = "ec")]
+        #[cfg_attr(doc_cfg, doc(cfg(feature = "ec")))]
         impl<$P> Sample<Standard> for $affine
         where
             $P: $trait,
@@ -48,6 +54,7 @@ macro_rules! sample_curve {
     };
 }
 
+#[cfg(feature = "ec")]
 sample_curve!(
     P,
     SWModelParameters,
@@ -55,6 +62,7 @@ sample_curve!(
     short_weierstrass_jacobian::GroupAffine<P>,
 );
 
+#[cfg(feature = "ec")]
 sample_curve!(
     P,
     TEModelParameters,
@@ -63,9 +71,12 @@ sample_curve!(
 );
 
 /// Builds a [`Sample`] implementation for all the `$fp` types.
+#[cfg(feature = "ff")]
 macro_rules! sample_fp {
     ($($fp:tt),* $(,)?) => {
         $(
+            #[cfg(feature = "ff")]
+            #[cfg_attr(doc_cfg, doc(cfg(feature = "ff")))]
             impl<P> Sample<Standard> for $fp<P>
             where
                 $fp<P>: UniformRand,
@@ -83,4 +94,5 @@ macro_rules! sample_fp {
     };
 }
 
+#[cfg(feature = "ff")]
 sample_fp!(Fp64, Fp256, Fp320, Fp384, Fp448, Fp768, Fp832);
